refactor(server): dissolve misc_routes.rs into cohesive route modules (GH-1797)
Split the 1120-line http/misc_routes.rs into focused route modules:

- health_routes.rs: health_check, project_queue_stats, reset_runtime_circuit_breaker
- workflow_routes.rs: issue/project workflow REST lookups plus the runtime tree handler
- rpc_routes.rs: handle_rpc
- github_webhook_routes.rs (+ github_webhook_routes_tests.rs): webhook intake
- intake_status_routes.rs: /api/intake status and recent-dispatch helpers
- signal_routes.rs: /signals ingestion
- auth_routes.rs: password reset endpoint

The three near-identical workflow lookup handlers now share one generic response helper parameterized by entity label; response bodies are byte-identical to before. The runtime_workflow_* data-bag accessors stay as free functions in intake_status_routes.rs: moving them onto WorkflowInstance would push harness-workflow/src/runtime/model.rs (794 lines) past the U-16 800-line hard limit.

Routes now resolve directly to their modules; the misc_routes re-export block in http/mod.rs is deleted and all use sites import from the new modules. CLAUDE.md drops the stale U-16 exemption for misc_routes.rs.
