+++ crates/harness-server/src/http/signal_routes.rs
@@ -1,0 +1,142 @@
+use axum::{
+    body::Bytes,
+    extract::State,
+    http::{HeaderMap, StatusCode},
+    Json,
+};
+use serde_json::json;
+use std::sync::Arc;
+
+use super::state::AppState;
+
+#[derive(serde::Deserialize)]
+pub(crate) struct IngestSignalRequest {
+    pub(crate) source: String,
+    #[serde(default)]
+    pub(crate) severity: Option<harness_core::types::Severity>,
+    pub(crate) payload: serde_json::Value,
+}
+
+/// Infer severity from a GitHub webhook payload: CI failure → High, changes_requested → Medium.
+pub(crate) fn infer_github_severity(
+    payload: &serde_json::Value,
+) -> Option<harness_core::types::Severity> {
+    if let Some(obj) = payload.as_object() {
+        if let (Some(action), Some(check_run)) = (
+            obj.get("action").and_then(|v| v.as_str()),
+            obj.get("check_run"),
+        ) {
+            if action == "completed" {
+                let conclusion = check_run
+                    .get("conclusion")
+                    .and_then(|v| v.as_str())
+                    .unwrap_or("");
+                if conclusion == "failure" {
+                    return Some(harness_core::types::Severity::High);
+                }
+            }
+        }
+        if let Some(review) = obj.get("review") {
+            let state = review.get("state").and_then(|v| v.as_str()).unwrap_or("");
+            if state.eq_ignore_ascii_case("changes_requested") {
+                return Some(harness_core::types::Severity::Medium);
+            }
+        }
+    }
+    None
+}
+
+/// POST /signals — ingest an external signal (CI failure, review feedback, etc.).
+///
+/// Validates the `x-hub-signature-256` HMAC-SHA256 header using the configured
+/// `server.github_webhook_secret`. Rate-limited to 100 requests per source per minute.
+pub(crate) async fn ingest_signal(
+    State(state): State<Arc<AppState>>,
+    headers: HeaderMap,
+    body: Bytes,
+) -> (StatusCode, Json<serde_json::Value>) {
+    let secret = match super::github_intake_status::github_webhook_secret_for_request(
+        &state.core.server.config.server,
+    ) {
+        Err(_) => {
+            return (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                Json(json!({"error": "server.github_webhook_secret not configured"})),
+            )
+        }
+        Ok(secret) => secret,
+    };
+
+    let signature = match headers
+        .get("x-hub-signature-256")
+        .and_then(|v| v.to_str().ok())
+    {
+        Some(sig) => sig,
+        None => {
+            return (
+                StatusCode::UNAUTHORIZED,
+                Json(json!({"error": "missing header x-hub-signature-256"})),
+            )
+        }
+    };
+
+    if !crate::webhook::verify_github_signature(secret, signature, body.as_ref()) {
+        return (
+            StatusCode::UNAUTHORIZED,
+            Json(json!({"error": "invalid webhook signature"})),
+        );
+    }
+
+    let req: IngestSignalRequest = match serde_json::from_slice(&body) {
+        Ok(r) => r,
+        Err(e) => {
+            return (
+                StatusCode::BAD_REQUEST,
+                Json(json!({"error": format!("invalid payload: {e}")})),
+            )
+        }
+    };
+
+    if !state
+        .observability
+        .signal_rate_limiter
+        .check_and_increment(&req.source)
+    {
+        return (
+            StatusCode::TOO_MANY_REQUESTS,
+            Json(json!({"error": "rate limit exceeded: max 100 signals per minute per source"})),
+        );
+    }
+
+    let severity = req.severity.unwrap_or_else(|| {
+        if req.source == "github" {
+            infer_github_severity(&req.payload).unwrap_or(harness_core::types::Severity::Low)
+        } else {
+            harness_core::types::Severity::Low
+        }
+    });
+
+    let signal =
+        harness_core::types::ExternalSignal::new(req.source.clone(), severity, req.payload.clone());
+    let signal_id = signal.id.clone();
+
+    if let Err(e) = state.observability.events.log_external_signal(&signal) {
+        tracing::error!(source = %req.source, "failed to store external signal: {e}");
+        return (
+            StatusCode::INTERNAL_SERVER_ERROR,
+            Json(json!({"error": "failed to store signal"})),
+        );
+    }
+
+    tracing::info!(
+        source = %req.source,
+        severity = ?severity,
+        signal_id = %signal_id,
+        "external signal ingested"
+    );
+
+    (
+        StatusCode::OK,
+        Json(json!({"status": "accepted", "id": signal_id.as_str()})),
+    )
+}
